Split Pipeline in Render and Compute, Metal part
