+++ src/backend/metal/ComputePipelineMTL.h
@@ -1,0 +1,41 @@
+// Copyright 2017 The NXT Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef BACKEND_METAL_COMPUTEPIPELINEMTL_H_
+#define BACKEND_METAL_COMPUTEPIPELINEMTL_H_
+
+#include "backend/ComputePipeline.h"
+
+#import <Metal/Metal.h>
+
+namespace backend {
+namespace metal {
+
+    class ComputePipeline : public ComputePipelineBase {
+        public:
+            ComputePipeline(ComputePipelineBuilder* builder);
+            ~ComputePipeline();
+
+            void Encode(id<MTLComputeCommandEncoder> encoder);
+            MTLSize GetLocalWorkGroupSize() const;
+
+        private:
+            id<MTLComputePipelineState> mtlComputePipelineState = nil;
+            MTLSize localWorkgroupSize;
+    };
+
+}
+}
+
+#endif // BACKEND_METAL_COMPUTEPIPELINEMTL_H_
